Feat : Router useParams 및 아이템 클릭시 이동 추가
