+++ src/containers/commons/CardGrid/CardGrid.stories.jsx
@@ -1,13 +1,12 @@
 import React from "react";
-import CardResponsiveGrid from ".";
-import { colors } from "@/_shared";
+import CardGrid from ".";
 
 export default {
-  title: "containers/CardResponsiveGrid",
-  component: CardResponsiveGrid,
+  title: "containers/CardGrid",
+  component: CardGrid,
 };
 
-export const Default = (args) => <CardResponsiveGrid {...args} />;
+export const Default = (args) => <CardGrid {...args} />;
 
 Default.args = {
   theme: "light",
