+++ src/containers/commons/index.js
@@ -1,5 +1,4 @@
 export { default as CardGrid } from "./CardGrid";
-export { default as CardResponsiveGrid } from "./CardResponsiveGrid";
 export { default as CardSmallGrid } from "./CardSmallGrid";
 export { default as CommentList } from "./CommentList";
 export { default as FeedGrid } from "./FeedGrid";
